nfact_dr now works. Will need future work however to incorprate cluster etc but ready for merge
